refactor: remove Strangler Fig re-exports and move ProjectStatsCalculator to presenters
Delete 3 obsolete re-export shims (queue/reviewQueue, services/logService,
services/runtimeSettings) and update all 26 import sites to point directly
to their real modules in frameworks/. Move ProjectStatsCalculator from
interface-adapters/services/ to interface-adapters/presenters/ where it
belongs architecturally.
