+++ src/interface-adapters/presenters/projectStats.calculator.ts
