+++ src/main/websocket.ts
@@ -2,8 +2,8 @@ import type { FastifyInstance } from 'fastify';
 import type { WebSocket } from 'ws';
 import type { ReviewProgress, ProgressEvent } from '../entities/progress/progress.type.js';
 import type { Dependencies } from './dependencies.js';
-import { getJobsStatus, setProgressChangeCallback, setStateChangeCallback, updateJobProgress } from '../queue/reviewQueue.js';
-import { onLog, type LogEntry } from '../services/logService.js';
+import { getJobsStatus, setProgressChangeCallback, setStateChangeCallback, updateJobProgress } from '../frameworks/queue/pQueueAdapter.js';
+import { onLog, type LogEntry } from '../frameworks/logging/logBuffer.js';
 
 const wsClients = new Set<WebSocket>();
