--- src/queue/reviewQueue.ts
@@ -1,26 +1,0 @@
-// Strangler Fig: Re-export from new location
-// This file will be removed once all imports are updated
-export {
-  initQueue,
-  getQueue,
-  shouldDeduplicate,
-  markJobProcessed,
-  clearJobDeduplication,
-  createJobId,
-  enqueueReview,
-  cancelJob,
-  getJobStatus,
-  getQueueStats,
-  getJobsStatus,
-  updateJobProgress,
-  setProgressChangeCallback,
-  setStateChangeCallback,
-  getJobProgress,
-} from '../frameworks/queue/pQueueAdapter.js';
-
-export type {
-  ReviewJob,
-  JobStatus,
-  ProgressChangeCallback,
-  StateChangeCallback,
-} from '../frameworks/queue/pQueueAdapter.js';
