--- src/services/runtimeSettings.ts
@@ -1,9 +1,0 @@
-// Strangler Fig: Re-export from new location
-// This file will be removed once all imports are updated
-export {
-  getModel,
-  setModel,
-  getSettings,
-} from '../frameworks/settings/runtimeSettings.js';
-
-export type { ClaudeModel } from '../frameworks/settings/runtimeSettings.js';
