+++ src/usecases/triggerReview.usecase.ts
@@ -1,5 +1,5 @@
 import type { Logger } from 'pino';
-import type { ReviewJob } from '../queue/reviewQueue.js';
+import type { ReviewJob } from '../frameworks/queue/pQueueAdapter.js';
 import type { ReviewRequestTrackingGateway, Platform } from '../interface-adapters/gateways/reviewRequestTracking.gateway.js';
 
 export interface TriggerReviewParams {
